[WebProfiler] [HttpKernel] profile redirections
closes #17501.

The profiler stores the current request attributes in a
current session when a `RedirectionResponse` is returned.
So the next request profile will inherit the previous request
attributes.

The main profiler layout displays a shortcut to a previous
redirection profile, along with some useful informations.

The web debug toolbar shows a notifying icon, meaning a shortcut
to a redirection profile is available in the request toolbar panel.